+++ Frock-backend/routes/Interface/REST/Resources/CreateFullRouteResource.cs
@@ -3,7 +3,7 @@
     public record CreateFullRouteResource
     (
         int Frequency,
-        int Price,
+        double Price,
         int Duration,
         List<int> StopsIds, // List of stop IDs that this route will include
         List<CreateScheduleResource> Schedules // List of schedules for the route
